Implement Tables.schema for JSONTables.Table
Fixes #17. This PR updates the `jsontable` constructor function that
returns a `JSONTables.Table` object. It now does stricter checks that
the incoming object or array is actually a table (strict object of
arrays or array of objects); it also checks the names and types that
will be expected, with particular consideration for the array of
objects, where the objects may not all have the same "schema". We do a
pass over the array of objects, "union"-ing the columns from across all
rows, along with the value types. This obviously adds a performance hit
for initial table construction, but ultimately leads to a much more
predictable user experience. If the cost is indeed too prohibitive, we
can address the potential of adding a fast-path constructor later on,
which shouldn't be hard or disruptive.